fix(header): some burger menu issues
close menue when any page changing item is clicked + if show all link has custom test send that to
tracking

HEYUI-358
