+++ packages/header/src/components/SubNav.tsx
@@ -4,7 +4,7 @@ import Grid from '@heycar-uikit/grid';
 import Typography from '@heycar-uikit/typography';
 
 import { DEFAULT_LOCALE } from '../constants/Header.constants';
-import { itemOnClick } from '../utils/headerItemHelpers';
+import { useNavigationItem } from '../hooks/useNavigationItem';
 import { getSubNavGroupDetails } from '../utils/navigationHelpers';
 
 import { SubNavProps } from './Navigation.types';
@@ -26,6 +26,8 @@ const SubNav = React.forwardRef<HTMLDivElement, SubNavProps>(
     ref,
   ) => {
     const groupDeets = getSubNavGroupDetails(subNavGroups, isDropDownMenu);
+
+    const { itemOnClick } = useNavigationItem(navItemId);
 
     return (
       <Grid.Row
@@ -114,7 +116,9 @@ const SubNav = React.forwardRef<HTMLDivElement, SubNavProps>(
                     itemOnClick({
                       fn: trackingFn,
                       obj: {
-                        label: `${subNavGroup.heading} show all link`,
+                        label: `${subNavGroup.heading} ${
+                          showAllLink.label || locale.showAllLabel
+                        }`,
                         parentLabel: navItemName,
                         type: 'subnav_item',
                         href: showAllLink.href,
